Move annotations induction functions
